rename(destinations): 위치 일급 컬렉션 destinations 패키지 이동
