+++ src/main/java/janggi/domain/strategy/CannonMoveStrategy.java
@@ -1,9 +1,9 @@
 package janggi.domain.strategy;
 
 import janggi.domain.board.BoardInfo;
+import janggi.domain.board.Destinations;
 import janggi.domain.board.Direction;
 import janggi.domain.board.Position;
-import janggi.domain.route.Destinations;
 import java.util.List;
 
 public class CannonMoveStrategy extends PieceStrategy {
