+++ src/main/java/janggi/domain/strategy/PalaceMoveStrategy.java
@@ -1,10 +1,10 @@
 package janggi.domain.strategy;
 
 import janggi.domain.board.BoardInfo;
+import janggi.domain.board.Destinations;
 import janggi.domain.board.Direction;
 import janggi.domain.board.PalacePosition;
 import janggi.domain.board.Position;
-import janggi.domain.route.Destinations;
 import java.util.List;
 
 public class PalaceMoveStrategy extends PieceStrategy {
